requirements.txt
fastapi
uvicorn
apify-client
python-dotenv
pydantic
redis
pymongo
qdrant-client
openai
geopy
